+++ CHANGELOG.md
@@ -21,7 +21,7 @@
 - 增加 `SignAuthUtils` 工具类
 - `zc-feign-spring-boot-starter` 远程通信通用组件，默认配置中增加SignAuth的参数，详细见`com.zhengcheng.feign.FeignAutoConfiguration`,其中默认的秘钥为：security.api.key，请在配置文件中写入秘钥，默认为 `zhengcheng` 
 - 技术分享 [API接口防止参数篡改和重放攻击](https://note.youdao.com/ynoteshare1/index.html?id=ed15f29e7ad1ff2d15a1236231283bc7&type=note)
-
+- 删除 `FeignInterceptorConfig` , 默认配置中增加 Feign OAuth2 拦截器
 
 ## [v4.3.0] 2020.4.18